crawler/requirements.txt
beautifulsoup4
requests==2.26.0
pandas
cryptography
config
cassandra-driver
